solana_tpu_client::nonblocking

Module tpu_client

Re-exports

pub use crate::tpu_client::Result;

Structs

LeaderTpuService

Service that tracks upcoming leaders and maintains an up-to-date mapping of leader id to TPU socket address.

TpuClient

Client which sends transactions directly to the current leader’s TPU port over UDP. The client uses RPC to determine the current leader and fetch node contact info

Enums

TpuSenderError