solana_tpu_client

Module tpu_client

Re-exports

pub use crate::nonblocking::tpu_client::TpuSenderError;

Structs

TpuClient

Client which sends transactions directly to the current leader’s TPU port over UDP. The client uses RPC to determine the current leader and fetch node contact info

TpuClientConfig

Config params for TpuClient

Constants

DEFAULT_FANOUT_SLOTS

Default number of slots used to build TPU socket fanout set

DEFAULT_TPU_CONNECTION_POOL_SIZE

The default connection count is set to 1 – it should be sufficient for most use cases. Validators can use –tpu-connection-pool-size to override this default value.

DEFAULT_TPU_ENABLE_UDP

DEFAULT_TPU_USE_QUIC

DEFAULT_VOTE_USE_QUIC

MAX_FANOUT_SLOTS

Maximum number of slots used to build TPU socket fanout set

Type Aliases

Result